% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_birdRT.R
\docType{data}
\name{birdRT}
\alias{birdRT}
\title{Bird Reaction Time Data}
\format{
A data frame with variables:
\describe{
\item{Condition}{Experimental condition}
\item{RT}{Reaction time}
}
}
\source{
Instructional dataset.
}
\usage{
birdRT
}
\description{
A small instructional dataset containing reaction time data by condition.  Hypothetical data were
inspired by Hahner, L., & Nieder, A. (2023). Costs and benefits of voluntary attention in crows. Royal Society Open Science, 10(8), 230517. https://doi.org/10.1098/rsos.230517
}
\keyword{datasets}
