% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{run_ez_anova}
\alias{run_ez_anova}
\title{ezANOVA wrapper for Essentials tools}
\usage{
run_ez_anova(...)
}
\arguments{
\item{...}{Passed to ez::ezANOVA().}
}
\value{
The result from ez::ezANOVA().
}
\description{
Convenience wrapper around ez::ezANOVA() used in the book.
}
