% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{run_pwr_t_test}
\alias{run_pwr_t_test}
\title{pwr.t.test wrapper for Essentials tools}
\usage{
run_pwr_t_test(...)
}
\arguments{
\item{...}{Passed to pwr::pwr.t.test().}
}
\value{
The result from pwr::pwr.t.test().
}
\description{
Convenience wrapper around pwr::pwr.t.test().
}
\examples{
run_pwr_t_test(n = NULL,
               d = 0.5,
               sig.level = 0.05,
               power = 0.8,
               type = "one.sample")

}
