% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{run_wp_ttest}
\alias{run_wp_ttest}
\title{webpower t-test wrapper for Essentials tools}
\usage{
run_wp_ttest(...)
}
\arguments{
\item{...}{Passed to \code{\link[WebPower:wp.t]{WebPower::wp.t()}}.}
}
\value{
The result from \code{\link[WebPower:wp.t]{WebPower::wp.t()}}.
}
\description{
Convenience wrapper around \code{\link[WebPower:wp.t]{WebPower::wp.t()}} used in the book.
}
\examples{
run_wp_ttest(n1 = 50,
             d = 0.5,
             alpha = 0.05,
             power = NULL,
             type = "one.sample")

}
