% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SleepData.R
\docType{data}
\name{SleepData}
\alias{SleepData}
\title{Sleep Duration Data}
\format{
A data frame with 26 rows and 5 variables:
\describe{
\item{Participant_ID}{Participant identifier.}
\item{Sleep_Duration}{Sleep duration.}
\item{Sex}{Participant sex.}
\item{Age}{Participant age in years.}
\item{Country}{Participant country.}
}
}
\source{
Dataset adapted for instructional use and modeled after
Chaumon and colleagues (2022). The Blursday database as a resource
to study subjective temporalities during COVID-19.
Nature Human Behaviour, 6(11), 1587-1599.
\doi{10.1038/s41562-022-01419-2}
}
\usage{
SleepData
}
\description{
A small instructional dataset containing sleep duration and
demographic information. The dataset is modeled after research on
subjective temporalities during COVID-19.
}
\keyword{datasets}
