% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birdRT}
\alias{birdRT}
\title{Bird reaction time dataset}
\format{
A data frame with the following columns:
\describe{
\item{Subject}{Participant identifier (character).}
\item{Condition}{Experimental condition (factor).}
\item{ReactionTime}{Reaction time (numeric).}
}
}
\source{
Imported from D:/data.CSV (original file location on author's machine).
}
\usage{
birdRT
}
\description{
A dataset used in the Essentials stats book chapter 3.
}
\examples{
head(birdRT)
}
\keyword{datasets}
